Folder structure was created. User DAO, DTO, Service, Controller, Exceptions and ExceptionsHandler were created. User main functionality was created and tested, though User is not done yet.

+++ src/main/java/com/example/Cookio/models/RecipeIngredient.java
@@ -4,6 +4,8 @@ import com.example.Cookio.helpers.RecipeIngredientId;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
@@ -26,8 +28,9 @@ public class RecipeIngredient {
     private Ingredient ingredient;
 
     @Column(name = "quantity", precision = 10, scale = 2)
-    private double quantity;
+    private BigDecimal quantity;
 
     @Column(name = "unit", length = 50)
     private String unit;
+
 }
